Add files via upload
This demonstrates how to use DLL's in C# in Visual Studio, and it also serves as a Proof of Concept for the upcoming Jimusho Dojo Management suite.  As the software evolves, it will contain enhanced features that will help Dojo owners/managers with the administrative tasks of running a Martial Arts facility.  The word Jimusho comes from the Japanese word for an administrative office where clerical work might be conducted, so it seemed appropriate for this project.